% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishobspgr-package.R
\docType{package}
\name{fishobspgr-package}
\alias{fishobspgr}
\alias{fishobspgr-package}
\title{fishobspgr: Client for BC Fish Observation and Passage RESTful API}
\description{
Client for BC Fish Observation and Passage RESTful API.
}
\author{
\strong{Maintainer}: Seb Dalgarno \email{seb.dalgarno@gmail.com} (\href{https://orcid.org/0000-0002-3658-4517}{ORCID})

}
\keyword{internal}
