% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection-properties.R
\name{fop_collection_passage_properties}
\alias{fop_collection_passage_properties}
\title{BC Fish Passage collection Properties}
\usage{
fop_collection_passage_properties(collection_id)
}
\arguments{
\item{collection_id}{A character string of the collection id.}
}
\value{
A tibble.
}
\description{
Get information about a fish passage collection's column names.
}
\examples{
collection_id <- "bcfishpass.barriers_falls"
fop_collection_passage_properties(collection_id)
}
\seealso{
Other collections: 
\code{\link{fop_collection_observation_properties}()},
\code{\link{fop_collections_observation}()},
\code{\link{fop_collections_passage}()},
\code{\link{fop_query_collection_observation}()},
\code{\link{fop_query_collection_passage}()}
}
\concept{collections}
