% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{fop_collections_passage}
\alias{fop_collections_passage}
\title{BC Fish Passage collections}
\usage{
fop_collections_passage()
}
\value{
A tibble.
}
\description{
Get information about the BC Fish Passage collections.
}
\examples{
fop_collections_passage()
}
\seealso{
Other collections: 
\code{\link{fop_collection_observation_properties}()},
\code{\link{fop_collection_passage_properties}()},
\code{\link{fop_collections_observation}()},
\code{\link{fop_query_collection_observation}()},
\code{\link{fop_query_collection_passage}()}
}
\concept{collections}
