% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{fop_query_collection_passage}
\alias{fop_query_collection_passage}
\title{Query BC Fish Observation Collection}
\usage{
fop_query_collection_passage(
  collection_id,
  filter = NULL,
  limit = 10000,
  offset = 0,
  bbox = NULL,
  properties = NULL,
  transform = NULL,
  sortby = NULL,
  groupby = NULL
)
}
\arguments{
\item{collection_id}{A character string of the collection id.}

\item{filter}{A named vector or list of the filter(s) to apply,
where the list names correspond to column names and the
list values correspond to the desired
value, e.g. \code{list(gnis_name = "Sangan River")}.}

\item{limit}{A positive whole number that limits the number of
features in the response.}

\item{offset}{A positive whole number that starts the response at an offset.}

\item{bbox}{A vector of four numbers indicating bounding box in lon/lat
(e.g. \code{c(minLon, minLat, maxLon, maxLat)}) that features in response
must intersect.}

\item{properties}{A vector of strings of the column names to include.
If NULL (default), all columns are retained.}

\item{transform}{A character vector with the name of the
\href{https://github.com/CrunchyData/pg_featureserv/blob/master/config/pg_featureserv.toml.example#L29}{\emph{valid transform}}
function followed by the parameter values (e.g. c("ST_Simplify", 100)).}

\item{sortby}{A string that sorts the response items by a property.
Default is ascending but property name can be prepended with
'-' to indicate descending.}

\item{groupby}{A string of the property name (i.e., column name) to group
response features by. This would typically be used with an aggregate
transform function.}
}
\description{
Query a collection from the BC Fish Observation API.
}
\examples{
collection_id <- "bcfishpass.barriers_falls"
filter <- list(barrier_type = "FALLS")
fop_query_collection_passage(collection_id, filter = filter)
}
\seealso{
Other collections: 
\code{\link{fop_query_collection_observation}()}
}
\concept{collections}
