% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{fop_collections_observation}
\alias{fop_collections_observation}
\title{BC Fish Observation collections}
\usage{
fop_collections_observation()
}
\value{
A tibble.
}
\description{
Get information about the BC Fish Observation collections.
}
\examples{
fop_collections_observation()
}
\seealso{
Other collections: 
\code{\link{fop_collections_passage}()}
}
\concept{collections}
